+++ app/src/app/triangular/triangular.module.js
@@ -5,6 +5,6 @@
         .module('triangular', [
             'ngMaterial',
             'triangular.layouts', 'triangular.components', 'triangular.themes', 'triangular.profiler', 'triangular.directives',
-            'ui.router',
+            'ui.router'
         ]);
 })();