+++ app/src/app/triangular/layouts/layouts.provider.js
@@ -7,27 +7,44 @@
 
     /* @ngInject */
     function layoutProvider() {
-        // Provider
-        var layout = {
-            sideMenuSize: 'full'
+        var layoutDefaults = {
+            toolbarSize: 'default',
+            toolbarShrink: true,
+            toolbarClass: '',
+            contentClass: '',
+            sideMenuSize: 'full',
+            footer: true
         };
+        var layout = {};
 
-        this.getOption = getOption;
-        this.setOption = setOption;
+        this.getDefaultOption = getDefaultOption;
+        this.setDefaultOption = setDefaultOption;
 
-        function getOption(name) {
-            return layout[name];
+        function getDefaultOption(name) {
+            return layoutDefaults[name];
         }
 
-        function setOption(name, value) {
-            layout[name] = value;
+        function setDefaultOption(name, value) {
+            layoutDefaults[name] = value;
         }
+
+        // init
+
+        angular.extend(layout, layoutDefaults);
 
         // Service
-        this.$get = function() {
+        this.$get = function($rootScope) {
+            $rootScope.$on('$stateChangeStart', function(event, toState) {
+                var layoutOverrides = angular.isDefined(toState.data) && angular.isDefined(toState.data.layout) ? toState.data.layout : {};
+                angular.extend(layout, layoutDefaults, layoutOverrides);
+            });
+
+            function setOption(name, value) {
+                layout[name] = value;
+            }
+
             return {
                 layout: layout,
-                getOption: getOption,
                 setOption: setOption
             };
         };