+++ app/src/app/triangular/triangular.module.js
@@ -1,0 +1,11 @@
+(function() {
+    'use strict';
+
+    angular
+        .module('triangular', [
+            'ngMaterial',
+            'triangular.layouts', 'triangular.components', 'triangular.themes', 'triangular.directives',
+            'ui.router',
+            'triangular.profiler'
+        ]);
+})();