+++ app/changelog.md
@@ -131,13 +131,12 @@
 - More modular structure
 - Authentication Module: Error messages now only show when input has been changed
 - Authentication Module: All pages now use site wide logo and name
-
 - Glup: Replaced JSHint with ESLint and ESLint Angular Plugin
 - Gulp: Stopped debug server crash when js has syntax error
 - Gulp: Added server port option
-
+- Gulp: Added eslint-angular plugin to check for AngularJS best practices
 - Breadcrumbs: New faster breadcrumbs
-
 - Layouts: new page layout examples
 - Layouts: new layout provider & service to easily modify page layouts
 - Dashboards: removed title attribute to stop popup title.
+- NEW Profiler module to check speed of AngularJS