+++ app/changelog.md
@@ -121,3 +121,22 @@
 > Release Date : ???
 
 - Fixed state in toolbars demo page (menu now ignores params on route)
+
+
+# Version 2.0.0
+> Release Date : ????
+
+- Code style changed to match https://github.com/johnpapa/angular-styleguide
+- Triangular moved into a module to make it easier to drop the code into existing apps or creating new ones
+- More modular structure
+- Authentication Module: Error messages now only show when input has been changed
+- Authentication Module: All pages now use site wide logo and name
+- Glup: Replaced JSHint with ESLint and ESLint Angular Plugin
+- Gulp: Stopped debug server crash when js has syntax error
+- Gulp: Added server port option
+- Gulp: Added eslint-angular plugin to check for AngularJS best practices
+- Breadcrumbs: New faster breadcrumbs
+- Layouts: new page layout examples
+- Layouts: new layout provider & service to easily modify page layouts
+- Dashboards: removed title attribute to stop popup title.
+- NEW Profiler module to check speed of AngularJS